// extensions/cg-raycast/src/types/cg/priceChange.ts
export interface IPriceChange {
  h12PriceChangePercent: number;
  h1PriceChangePercent: number;
  h24PriceChangePercent: number;
  h4PriceChangePercent: number;
  m15PriceChangePercent: number;
  m30PriceChangePercent: number;
  m5PriceChangePercent: number;
  price: number;
  symbol: string;
  symbolLogo: string;
  volUsd: number;
}